Rename class(typo)
