+++ app/src/main/kotlin/dvachmovie/fragment/start/StartFragment.kt
@@ -5,8 +5,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.lifecycle.Observer
-import androidx.navigation.fragment.findNavController
-import dvachmovie.Constraints
+import dvachmovie.Constants
 import dvachmovie.architecture.base.BaseFragment
 import dvachmovie.databinding.FragmentStartBinding
 import dvachmovie.di.core.FragmentComponent
@@ -45,7 +44,7 @@ class StartFragment : BaseFragment<StartVM,
 
     private fun prepareData() {
         movieRepository.observe(this, Observer { movies ->
-            if (settingsUseCase.getLoadingParam() == Constraints.LOADING_EVERY_TIME ||
+            if (settingsUseCase.getLoadingParam() == Constants.LOADING_EVERY_TIME ||
                     movies.size < 100) {
                 dvachUseCase.getNumThreads("b", initWebm())
             } else {
