+++ data/src/main/java/dvachmovie/api/RetrofitSingleton.kt
@@ -1,6 +1,6 @@
 package dvachmovie.api
 
-import dvachmovie.Constraints.Companion.BASE_URL
+import dvachmovie.Constants.Companion.BASE_URL
 import okhttp3.HttpUrl
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
